third_party/WebKit/LayoutTests/http/tests/devtools/sources/inline-script-with-source-map-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests inline source with source map.
Set breakpoint at console.log line
Call function and dump stack trace
Script execution paused.
Call stack:
    0) foo (a.ts:2)
    1)  (:1)
Dump console mesage with its location:
Script execution resumed.
a.ts:2 42
